Finish entity reorganization
I still do not feel it is there yet, but it is a bit better
