+++ core/src/db/entity/book_club/club.rs
@@ -8,9 +8,11 @@ use crate::{
 };
 
 use super::{
-	book_club_member_and_schedule_include, book_club_member_user_username,
-	book_club_with_books_include, BookClubMember, BookClubMemberRoleSpec,
-	BookClubSchedule,
+	prisma_macros::{
+		book_club_member_and_schedule_include, book_club_member_user_username,
+		book_club_with_books_include,
+	},
+	BookClubMember, BookClubMemberRoleSpec, BookClubSchedule,
 };
 
 #[derive(Default, Debug, Clone, Deserialize, Serialize, Type, ToSchema)]
