+++ core/src/db/entity/book_club/common.rs
@@ -1,0 +1,1 @@
+
