+++ core/src/db/entity/book_club/schedule.rs
@@ -7,7 +7,7 @@ use crate::{
 	prisma::{book_club_book, book_club_schedule},
 };
 
-use super::{book_club_with_books_include, BookClubChatBoard};
+use super::{prisma_macros::book_club_with_books_include, BookClubChatBoard};
 
 #[derive(Default, Debug, Clone, Deserialize, Serialize, Type, ToSchema)]
 pub struct BookClubSchedule {
