+++ core/src/db/entity/mod.rs
@@ -20,7 +20,13 @@ pub use media::*;
 pub use metadata::*;
 pub use reading_list::*;
 pub use series::*;
+pub use server_config::*;
 pub use tag::*;
 pub use user::*;
 
-pub use common::{FileStatus, LayoutMode};
+pub use common::{Cursor, FileStatus, LayoutMode};
+
+pub mod macros {
+	pub use super::book_club::prisma_macros::*;
+	pub use super::metadata::prisma_macros::*;
+}
